# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получение списка задач",
                "operationId": "573d829b0171575d1b3e842dbe53fe52",
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TaskResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создание новой задачи",
                "operationId": "1d9aacd182a029b7dab850a73fe0d481",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/{task}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удаление задачи",
                "operationId": "93810fac7ceb4a4cfcfe7e5a33df7c3c",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Задача успешно удалена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновление задачи",
                "operationId": "d51aff711c6c2c7776b6ef92997cb559",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tasks/{task}/complete": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Отметить задачу как выполненную",
                "operationId": "0e4a68d0b876778af9e7a5a6a7487785",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача отмечена как выполненная",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TaskRequest": {
                "title": "TaskRequest",
                "description": "Запрос на создание или обновление задачи",
                "required": [
                    "title",
                    "priority",
                    "category"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Сделать тест"
                    },
                    "description": {
                        "type": "string",
                        "example": "Подготовить отчет",
                        "nullable": true
                    },
                    "priority": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high"
                        ],
                        "example": "high"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "work",
                            "study",
                            "personal"
                        ],
                        "example": "study"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30 12:00:00",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TaskResource": {
                "title": "TaskResource",
                "description": "Ресурс задачи",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Сделать тест"
                    },
                    "description": {
                        "type": "string",
                        "example": "Подготовить отчет"
                    },
                    "priority": {
                        "type": "string",
                        "example": "high"
                    },
                    "category": {
                        "type": "string",
                        "example": "study"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-30 12:00:00"
                    },
                    "completed": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        }
    },
...
}